refactor banner
