+++ frontend/src/components/molecules/Banner.jsx
@@ -2,7 +2,11 @@
 import styled from 'styled-components'
 import PropTypes from 'prop-types'
 import { colors, device } from '../../theme'
-import { BannerStudents, BannerBcn, BannerBarcelona } from '../atoms/Contributors'
+import ColaboradoresBanner from '../../assets/logos/Colaboradores-Banner.jpg'
+import BarcelonaBanner from '../../assets/logos/barcelona-banner.jpg'
+import Contributors from '../organisms/Contributors'
+// eslint-disable-next-line import/no-cycle
+import { Text } from '../atoms'
 
 const BannerStyled = styled.section`
   display: flex;
@@ -19,17 +23,58 @@ const BannerStyled = styled.section`
     padding: 0 8rem 0 8rem;
   }
 
-  p {
+  .colaboradores-text {
+    margin-top: 25rem;
+  }
+  .contributors-avatar {
+    margin-top: 2rem;
+  }
+  .footer {
+    margin-bottom: 2rem;
+  }
+  .directorio {
+    margin-top: 25rem;
+  }
+  h4 {
     font-size: 1.5rem; /* 24px */
     line-height: 2rem; /* 32px */
     font-weight: bold;
+    margin-bottom: 0;
     color: ${colors.redPink};
   }
 `
+const Img = styled.img`
+  position: absolute;
+  left: 0;
+  height: 23rem;
+  width: 100%;
+  object-fit: cover;
+  object-position: bottom;
+`
 
 function Banner({ students }) {
-  // return <BannerStyled>{students ? <BannerStudents /> : <BannerBcn />}</BannerStyled>
-  return <BannerStyled>{students ? <BannerBarcelona /> : <BannerBcn />}</BannerStyled>
+  const srcImg = students ? ColaboradoresBanner : BarcelonaBanner
+
+  return (
+    <BannerStyled alignItems="center">
+      <Img src={srcImg} alt="banner image" />
+      {students ? (
+        <>
+          <Text as="h4" className="colaboradores-text" text="Colaboradores" />
+          <Text text="Usuarios de github.com que han contribuido" />
+          <Contributors />
+          <Text as="h4" text="Perfiles" />
+          <Text className="footer" text="ReactJS y Node Juniors Developers en Barcelona" />
+        </>
+      ) : (
+        <Text
+          as="h4"
+          className="directorio"
+          text="Un directorio abierto desarrollado por los alumnos de Barcelona Activa"
+        />
+      )}
+    </BannerStyled>
+  )
 }
 
 Banner.propTypes = {
